PWGEM/PhotonMeson: adjust naming convention
