+++ PWGEM/PhotonMeson/Tasks/CMakeLists.txt
@@ -34,13 +34,13 @@ o2physics_add_dpl_workflow(pcm-qc-mc
                     PUBLIC_LINK_LIBRARIES O2::Framework O2::DetectorsBase O2Physics::AnalysisCore O2Physics::PWGEMPhotonMesonCore
                     COMPONENT_NAME Analysis)
 
-o2physics_add_dpl_workflow(dalitz-qc
-                    SOURCES dalitzQC.cxx
+o2physics_add_dpl_workflow(dalitz-ee-qc
+                    SOURCES dalitzEEQC.cxx
                     PUBLIC_LINK_LIBRARIES O2::Framework O2::DetectorsBase O2Physics::AnalysisCore O2Physics::PWGEMPhotonMesonCore
                     COMPONENT_NAME Analysis)
 
-o2physics_add_dpl_workflow(dalitz-qc-mc
-                    SOURCES dalitzQCMC.cxx
+o2physics_add_dpl_workflow(dalitz-ee-qc-mc
+                    SOURCES dalitzEEQCMC.cxx
                     PUBLIC_LINK_LIBRARIES O2::Framework O2::DetectorsBase O2Physics::AnalysisCore O2Physics::PWGEMPhotonMesonCore
                     COMPONENT_NAME Analysis)
